Fix the directory name of the new addons (#1211)
For a UserJS addon, the expected script name will be the same as the
directory with a suffix ".js". So `AddonsContentsWidget::addAddon()`
must only remove the suffix of the selected addon file, using Qt's
`completeBaseName()` instead of truncating at the *first dot* with
`baseName()`.

The previous behavior made addons like `x.user.js` saved under
`x/x.user.js`, and then not loaded because `x/x.js` was expected.
Resolves #1210

In case the User Script file is not found, a message is logged.